Promote the time_display method from Comment to EntryLite to make things easier on layouts that use the same code for printing both entries and comments. The basic EntryLite implementation just prints out a simple date and time string with no timezone information.
Also added support for passing a date or time format of "none" to suppress the date or time. Kinda ugly, but was the best I could do while retaining back-compat. This is necessary for layouts that don't print the date and time all as one string.

Updated Quite Lickable to support time_display as an example.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@6871 1dfba125-df10-0410-8a77-af1a8d6d074f