Move the logic for whether to display the talkread and talkpost links into the S2 constructors. This allows layouts to avoid hardcoding this logic, and thus makes it easier to change in future.
Two new members are provided on CommentInfo which tell layer code whether to display
each link. Note that it is possible for the "enabled" member to be true but for both
of these to be false, such as on the talkpost page when there aren't yet any comments:
Comments *are* enabled, but the readlink isn't visible because there aren't any comments!

Quite Lickable is updated as an example. Due to the above, it now checks
(show_readlink or show_postlink) rather than (enabled), since it doesn't really
care whether comments are enabled or not, it just wants to know whether either of
the links are to be displayed.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@7097 1dfba125-df10-0410-8a77-af1a8d6d074f