Revert some of the changes made at http://community.livejournal.com/changelog/3012117.html
Specifically, change the overloaded function back to the way it was, so it doesn't break theme layers
that depend on that function.  This means that anyone who does overload the function won't get time
zone support (unless they update their version of the function), but that's okay.

I wanted to use "newnull EntryLite" instead of "new EntryLite" but S2 didn't recognize it for some reason.
The $e.depth > 0 check works just fine though, since a new EntryLite object sets all integers to 0 (according
to the docs, anyway, and testing confirms that it works).



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@7104 1dfba125-df10-0410-8a77-af1a8d6d074f