A few bugfixes for the S2 editor.
 * It now works sensibly in Opera. (Yay! Fanfare! etc)
 * Fixed the daft bug in IE where dragging the bars caused everything to get selected
 * Don't force a particular set of fonts for the editor. It was just forcing the defaults anyway.
 * The editor and reference pane no longer wrap. Wrapping source code makes it hard to read, and the reference pane just looked daft when the stuff inside was wrapped.

This is tested in IE for Windows, Mozilla and Opera. I don't have a Mac, so I've not tested in Safari or MacIE.
Given that this code is held together with gaffer tape and chewing gum, it's probably worth testing it on a Mac as it might well have gone crazy.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@7403 1dfba125-df10-0410-8a77-af1a8d6d074f