hbox/vbox support for A Sturdy Gesture and Bloggish.
For Bloggish, there's three themes where the header banner is the full width of the window, and in those cases, the hbox_top looks better
below it than above it.  So I special-cased those themes.

Also added in a check for if viewer_sees_hbox_bottom and for if we're viewing a recent or friends page to Bloggish's 5linkunit, since it wasn't
there before.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@7615 1dfba125-df10-0410-8a77-af1a8d6d074f