The Bloggish CSS and images are now hosted on LJ so we don't have to rely on TypePad.
We decided to put the files in ljcom and link directly to livejournal.com from Bloggish so that other sites
can still use the CSS, but can't modify it.  So basically, for other sites, it'll be just as before, except
they're getting the CSS from LJ now instead of TypePad.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8107 1dfba125-df10-0410-8a77-af1a8d6d074f