Stopgap solution to the bug where you'd see the wrong usernames if you pass in a username that starts with numbers: pass in the entire comm object, not just the username.
(Explanation: sub want_user gets passed a username like "70s", this passes the int() check and so it's treated as a userid, and we return the $u object for userid 70, not user "70s". Problem: There's no way to distinguish between the user "8674309" vs the userid "8675309" when we pass that in. If anyone has a good solution for this, please advise!)



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8717 1dfba125-df10-0410-8a77-af1a8d6d074f