Allow an S2 layer to be mapped to another layer.  In other words, two copies of a layer can be compiled and stored as separate without duplicating code.  All that needs to go in the new layer's *.s2 is the layerinfo lines that give it a name, des, etc.
This is useful for the case where we want to create two layouts that have the same DOM and feature set, but we want them to
have different themes.  And they'll be listed as separate layouts in the drop-down.  They actually are two separate layouts,
but one is made from the other, so only one copy of the code has to be maintained.

To map a layer:
In the s2layers*.dat file, the line with the new layer should be something like this:

new_layout/layout layout(layout_to_copy_from/layout) core1

Thanks to <lj user="hachi"> for help with figuring out how to do this.  He also committed a test that shows that a layout
with two sets of layerinfo definitions will always ignore the first set and use the second set.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8754 1dfba125-df10-0410-8a77-af1a8d6d074f