Fix the bug where:
* lj user tags in S2 get expanded in the title of the page, so you see HTML in the titlebar
* When there's no subject, talkread takes from the entry itself; if it starts with an lj user, HTML is shown in the title

Reason: the lj user spans weren't respecting the "textonly" opt in cleanhtml -- we were cleaning out anything of the form lj user="foo", but when that's expanded, it's a giant span. So we don't get caught in the section where we're checking for lj-specific tags, and we don't check for the textonly flag when we're looking at spans for ljuser.

Resolves two tickets:
   http://rt.livejournal.org/Ticket/Display.html?id=751
   http://rt.livejournal.org/Ticket/Display.html?id=1474

And we now pass the cleaner-subject test, yay.

Many thanks to Henry and Janine for the head-banging all throughout today; this has driven all of us nuts. :)



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8790 1dfba125-df10-0410-8a77-af1a8d6d074f