Copy cleanup for this email.
I took out the check for whether the new journal is a friend of the person getting the notif; presumably if the account was just created, I haven't had time to add it to my friends list yet.

Also, I couldn't figure out why we had the line to return if we don't have $u and $self->friend -- I don't see any other subscriptions that do this. Left 'em in for now, though, but I'm not sure we need them?



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8794 1dfba125-df10-0410-8a77-af1a8d6d074f