From imc:
In rev 9144, the regex match for deciding whether to un-utf8 the attributes of a tag was turned into a call to is_ascii. But is_ascii doesn't currently work on UTF-8 text.

The problem is that if you have, say &hearts; in a UTF-8 string, it
matches /[\x2665]/ and therefore doesn't match [\x80-\xff], so is_ascii
returns true and the cleaner doesn't bother to un-utf8 it.

This might work instead:

return ($text !~ m/[^\x01-\x7f]/);



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@9513 1dfba125-df10-0410-8a77-af1a8d6d074f