http://rt.livejournal.org/Ticket/Display.html?id=3142 http://rt.livejournal.org/Ticket/Display.html?id=3143
Return an error message if you're trying to add a new friend who's not visible (i.e. deleted, purged, or
suspended).  If you are modifying a non-visible friend, only allow removal of the friend, not modification
of friends groups or colors (this errors when you submit the form anyway).

This resolves the problem with trying to add a feed at /syn/ that's not visible as well.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@9544 1dfba125-df10-0410-8a77-af1a8d6d074f