[int: ads]
Changed this logic around a little.  Instead of having viewer_sees_ebox() only check the ebox
prop, I made it so the current_box_type() function will still return ebox if the prop is set
but no page is defined.  That way, the viewer_sees_ebox() function will work properly both
when a page is defined (it will check if the page supports eboxes) and when a page is not
defined (it'll ignore that and rely solely on the ebox prop setting).



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@9790 1dfba125-df10-0410-8a77-af1a8d6d074f