While working on the tag system tonight, I noticed this query:
SELECT journalid, jitemid, kwid FROM logtags WHERE ((journalid = '2' AND
jitemid = '11') OR (journalid = '2' AND jitemid = '7') OR (journalid = '2'
AND jitemid = '2') OR (journalid = '2' AND jitemid = '1') OR (journalid
= '2' AND jitemid = '16') OR (journalid = '2' AND jitemid = '13') OR
(journalid = '2' AND jitemid = '6') OR (journalid = '2' AND jitemid = '3')
OR (journalid = '2' AND jitemid = '12') OR (journalid = '2' AND jitemid
= '15') OR (journalid = '2' AND jitemid = '14') OR (journalid = '2' AND
jitemid = '8') OR (journalid = '2' AND jitemid = '4') OR (journalid = '2'
AND jitemid = '10') OR (journalid = '2' AND jitemid = '5'))

My soul cried a bit.  This patch refactors the code that generates that
query into something that looks more like this:

SELECT journalid, jitemid, kwid FROM logtags WHERE (journalid = 2
AND jitemid IN (11,7,2,1,16,13,6,3,12,15,14,8,4,10,5))

I also noticed that negative cacheing wasn't being used, so these queries
were running far more often than intended.  Bleh.

So:

    * fix queries to be written better
    * employ negative cacheing if there are no tags on an entry



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@10063 1dfba125-df10-0410-8a77-af1a8d6d074f