This patch fixes an outstanding issue in the tag system that caused it to generate obscenely large queries in the name of loading thousands of keywords used in tags.
The old query had WHERE (userid = ? AND kwid IN (N,N,N,N)) whereas the new
one simply does WHERE userid IN (N,N,N,N) which is far more efficient from
a database point of view.  Well, more efficient in the ways we care about!

The keywords are then stored in memcache, so this functionality can be used
by other parts of the site going forward.  (But to take full advantage
of that get_keyword_id should do the loading/memcacheing and not the tag
system.)

Cache is cleared every time a keyword is added.  Given that keywords are
never removed from the database, and they are only created through one API
(get_keyword_id), this works fine.




git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@10065 1dfba125-df10-0410-8a77-af1a8d6d074f