* add request cacheing to get_usertags
This is done because it's the best way to compensate for situations in
which tags get loaded twice.  This happens when the S2 perl modules load
the user's tags in order to render entries, and then the user's style calls
for tags to get loaded in order to render a tag sidebar or similar.

Previous to this patch, memcache was getting hit twice for the same
information in one request.  Respects the tag system's reset_cache sub and
is properly cleared in start_request.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@10289 1dfba125-df10-0410-8a77-af1a8d6d074f