Redo random user system.
* on post (public, not opted out, to personal journal) a user may end up
  with a row in the new random_user_set table

* clean_caches cron job will clean old entries from the table when run

* selecting a random user is done with LJ::User->load_random_user() and is
  done efficiently, will scale

How the system works, if a user wants to not appear in the random user
selection, they can use the latest_optout option at the console.  This
seemed to be the best prop to use.

Adjustable random period, by default 7 days is how big of a window we keep.
Can be changed in the configuration.




git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@10510 1dfba125-df10-0410-8a77-af1a8d6d074f