Add some lazy cleaning to community moderation. Specifically, there seem to be some cases (one every few weeks or so) where a community will have a modlog row but now modblob row; so they'll see "this entry is awaiting approval" but when they click through, it says "no entry, already handled?"
There's only two places where we delete from modlog/modblob (here and in the protocol), and the only thing I can think of is that this could happen in the event of a database unavailable, perhaps. (Where we'll delete from modblob, but not modlog, or vice versa.) Basically we'd need to ensure that *two* separate delete transactions happen, but it seems kind of messy to hang on to the old values of the first query and re-insert them if the second query fails.

So, lazy-clean: if we get a "no entry, already handled?", go ahead and delete that modlog. Doesn't matter if it doesn't exist, though, since this error could also be triggered if two moderators race each other.

Sanity checking (as always) apppreciated.


git-svn-id: http://code.livejournal.org/svn/livejournal/branches/r3@10559 1dfba125-df10-0410-8a77-af1a8d6d074f