Make it so we don't have to hard-code styling to position ads when they're in BML pages (besides, schemes often have different styling for different ads).
So basically, call ad_display if printing an ad directly from a BML page, and this will automatically
give a wrapper to the ad that will position it correctly.  Can also optionally pass a "below_ad" option
to put something below the ad.

Update uses of ad_display as well.



git-svn-id: http://code.livejournal.org/svn/livejournal/branches/r5@10993 1dfba125-df10-0410-8a77-af1a8d6d074f