Bring the "S1 as an S2 style" implementation of S1 up-to-date with all this crazy new HEAD stuff and ads and control strips and so forth, and add a configurable toggle to enable it for testing purposes.
When it's turned on, it'll render with both traditional S1 and this new version and check for any differences. It can optionally generate a diff
in the output to help locate the differences.

Thus far it seems to only differ in the contents of the HEAD element, because the S2 page constructors generate the elements there in a slightly
different order. However, I've not tested with ljcom's ads and navigation strip so I don't know if they're going to work properly.

The setting for enabling this new implementation is a coderef so it can conditionally enable S1w2 depending on the journal user, remote user or view.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@11915 1dfba125-df10-0410-8a77-af1a8d6d074f