Change password reset flow a bit. Right now it's:
lostinfo.bml -> click link in email -> get an email with a new password -> go to changepassword.bml and use that new password

Instead, this will change it to:

lostinfo.bml -> link in email authorizes you to use changepassword.bml without an old password.

Cuts out the step of the random password, which is by far the step that caused the most confusion and sturm und drang.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@12416 1dfba125-df10-0410-8a77-af1a8d6d074f