BugID:61069 -- Fix LJ::Worker::ProcessSub memory bloat.
I actually had to diagram this all out on a sheet of paper:

$u->{_notification_inbox}->{items}->[#]->{event}->{u} = $u;

alternately:

$u->{_inbox_items}->{#}->{event}->{u} = $u;

Change LJ::Event to use "userid" instead of storing the entire $u object. Only one caller directly into the hashref; the other two were 100% extraneous and never used elsewhere within the subroutine, so they can go.

How this is triggered, and why I didn't catch it the first time around: LJ::NotificationInbox->enqueue checks if you're over the max notifications in the inbox. If you are, it deletes the oldest ones that aren't bookmarks. To check if they're bookmarks, we had to load all the items in the inbox. When we do so, we instantiate LJ::Event objects. So because my tests never bumped into the notification limit, I didn't catch it in the first go-round.



git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@13002 1dfba125-df10-0410-8a77-af1a8d6d074f