LJSUP-2531
There's an autovivifying bug where a custom layout can end up in the public layers hash, but I can't figure out where this happens.
Instead, make it so LJ::Customize->get_current_theme won't fail if this bug occurs; it should assume the layer is custom unless it has a uniq defined in the hash.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@14076 1dfba125-df10-0410-8a77-af1a8d6d074f